fix(utils): fix review issues
