+++ convert_intro_to_md/utils/tei_renderer.py
@@ -1,6 +1,7 @@
 """Render a list of Block IR nodes to a TEI XML string."""
 
 import io
+import sys
 import xml.etree.ElementTree as ET
 
 from utils.nodes import (
@@ -37,8 +38,6 @@ _INLINE_WRAP: dict[type, tuple[str, str | None]] = {
     Paragraph: ("p", None),
 }
 
-# Module-level notes lookup; populated by render() before walking the tree.
-_notes_lookup: dict[int, Note] = {}
 
 
 def render(blocks: list[Block], intro_id: str, intro_locale: str) -> str:
@@ -59,19 +58,16 @@ def render(blocks: list[Block], intro_id: str, intro_locale: str) -> str:
     Returns:
         str: A serialized TEI XML string with an XML declaration, UTF-8 encoded.
     """
-    global _notes_lookup  # pylint: disable=global-statement
     ET.register_namespace("", _TEI_NS)
     ET.register_namespace("xml", _XML_NS)
 
-    _notes_lookup = _build_notes_lookup(blocks)
+    lookup = _build_notes_lookup(blocks)
 
     root = ET.Element(_tei("TEI"))
 
     _build_tei_header(root, intro_id, intro_locale)
 
-    _build_tei_body(root, blocks, intro_locale)
-
-    _notes_lookup = {}
+    _build_tei_body(root, blocks, intro_locale, lookup)
 
     ET.indent(root, space="  ")
     _fix_mixed_content_indent(root)
@@ -90,7 +86,9 @@ def _build_notes_lookup(blocks: list[Block]) -> dict[int, Note]:
 
     Iterates every block's ``notes`` list and indexes each note by the integer
     suffix of its ``id`` (e.g. ``'note-3'`` → ``3``).  Notes whose ``id``
-    cannot be parsed are silently skipped.
+    cannot be parsed are silently skipped.  When two notes share the same
+    integer suffix the first occurrence is kept and a warning is printed to
+    ``stderr``.
 
     Args:
         blocks (list[Block]): The IR blocks whose notes should be indexed.
@@ -102,9 +100,17 @@ def _build_notes_lookup(blocks: list[Block]) -> dict[int, Note]:
     for block in blocks:
         for note in block.notes:
             try:
-                lookup[int(note.id.split("-")[-1])] = note
+                n = int(note.id.split("-")[-1])
             except (ValueError, IndexError):
                 pass
+            else:
+                if n in lookup:
+                    print(
+                        f"Duplicate note number {n!r} encountered; keeping first.",
+                        file=sys.stderr,
+                    )
+                else:
+                    lookup[n] = note
     return lookup
 
 
@@ -129,7 +135,12 @@ def _build_tei_header(root: ET.Element, intro_id: str, intro_locale: str) -> Non
     ET.SubElement(lang_usage, _tei("language")).set("ident", intro_locale)
 
 
-def _build_tei_body(root: ET.Element, blocks: list[Block], intro_locale: str) -> None:
+def _build_tei_body(
+    root: ET.Element,
+    blocks: list[Block],
+    intro_locale: str,
+    lookup: dict[int, Note],
+) -> None:
     """Append a ``<text><body>`` subtree to *root*.
 
     Creates one ``<div>`` per block.  Each div receives an optional
@@ -139,6 +150,8 @@ def _build_tei_body(root: ET.Element, blocks: list[Block], intro_locale: str) ->
         root (ET.Element): The ``<TEI>`` root element.
         blocks (list[Block]): The IR blocks to render.
         intro_locale (str): Written as the ``xml:lang`` attribute on ``<text>``.
+        lookup (dict[int, Note]): Mapping from note number to Note, passed to
+            :func:`_render_node`.
     """
     text_el = ET.SubElement(root, _tei("text"))
     text_el.set(_xml("lang"), intro_locale)
@@ -151,9 +164,9 @@ def _build_tei_body(root: ET.Element, blocks: list[Block], intro_locale: str) ->
         if block.heading:
             head_el = ET.SubElement(div, _tei("head"))
             for node in block.heading:
-                _render_node(node, head_el)
+                _render_node(node, head_el, lookup)
         for node in block.content:
-            _render_node(node, div)
+            _render_node(node, div, lookup)
 
 
 # ---------------------------------------------------------------------------
@@ -161,32 +174,34 @@ def _build_tei_body(root: ET.Element, blocks: list[Block], intro_locale: str) ->
 # ---------------------------------------------------------------------------
 
 
-def _render_node(node: Node, parent: ET.Element) -> None:
+def _render_node(node: Node, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a single IR node as a child of *parent*.
 
     Args:
         node (Node): The IR node to render.
         parent (ET.Element): The ET element to attach rendered output to.
+        lookup (dict[int, Note]): Mapping from note number to Note, forwarded
+            to footnote and recursive renderers.
     """
     if isinstance(node, Text):
         _append_text(parent, node.value)
     elif isinstance(node, FootnoteRef):
-        _render_footnote_ref(node, parent)
+        _render_footnote_ref(node, parent, lookup)
     elif isinstance(node, CrossRef):
         _render_cross_ref(node, parent)
     elif isinstance(node, Ref):
-        _render_ref(node, parent)
+        _render_ref(node, parent, lookup)
     elif type(node) in _INLINE_WRAP:
-        _render_inline_node(node, parent)
+        _render_inline_node(node, parent, lookup)
     elif isinstance(node, Blockquote):
-        _render_blockquote(node, parent)
+        _render_blockquote(node, parent, lookup)
     elif isinstance(node, ListBlock):
-        _render_list_block(node, parent)
+        _render_list_block(node, parent, lookup)
     elif isinstance(node, Table):
-        _render_table(node, parent)
+        _render_table(node, parent, lookup)
 
 
-def _render_inline_node(node: Node, parent: ET.Element) -> None:
+def _render_inline_node(node: Node, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render an inline wrapper node from :data:`_INLINE_WRAP` as a child of *parent*.
 
     Looks up the TEI tag name and optional ``rend`` attribute from
@@ -196,27 +211,29 @@ def _render_inline_node(node: Node, parent: ET.Element) -> None:
     Args:
         node (Node): An inline IR node whose type is a key in :data:`_INLINE_WRAP`.
         parent (ET.Element): The ET element to attach the rendered element to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_node`.
     """
     tag, rend = _INLINE_WRAP[type(node)]
     el = ET.SubElement(parent, _tei(tag))
     if rend:
         el.set("rend", rend)
     for child in node.children:  # type: ignore[union-attr]
-        _render_node(child, el)
+        _render_node(child, el, lookup)
 
 
-def _render_blockquote(node: Blockquote, parent: ET.Element) -> None:
+def _render_blockquote(node: Blockquote, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a :class:`~utils.nodes.Blockquote` IR node as a TEI ``<quote>``.
 
     Args:
         node (Blockquote): The blockquote IR node.
         parent (ET.Element): The ET element to attach the ``<quote>`` to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_node`.
     """
     quote_el = ET.SubElement(parent, _tei("quote"))
     for para in node.paragraphs:
         el = ET.SubElement(quote_el, _tei("l"))
         for child in para.children:
-            _render_node(child, el)
+            _render_node(child, el, lookup)
 
 
 def _render_cross_ref(node: CrossRef, parent: ET.Element) -> None:
@@ -231,29 +248,34 @@ def _render_cross_ref(node: CrossRef, parent: ET.Element) -> None:
     el.text = str(node.n)
 
 
-def _render_footnote_ref(node: FootnoteRef, parent: ET.Element) -> None:
+def _render_footnote_ref(
+    node: FootnoteRef, parent: ET.Element, lookup: dict[int, Note]
+) -> None:
     """Render a :class:`~utils.nodes.FootnoteRef` IR node as a TEI ``<note>``.
 
     Args:
         node (FootnoteRef): The footnote-reference IR node.
         parent (ET.Element): The ET element to attach the ``<note>`` to.
+        lookup (dict[int, Note]): Mapping from note number to Note used to
+            populate the inline note content.
     """
     note_el = ET.SubElement(parent, _tei("note"))
     note_el.set("place", "end")
     note_el.set("n", str(node.n))
     note_el.set(_xml("id"), f"note-{node.n}")
-    note = _notes_lookup.get(node.n)
+    note = lookup.get(node.n)
     if note:
         for child in note.children:
-            _render_node(child, note_el)
+            _render_node(child, note_el, lookup)
 
 
-def _render_list_block(node: ListBlock, parent: ET.Element) -> None:
+def _render_list_block(node: ListBlock, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a :class:`~utils.nodes.ListBlock` IR node as a TEI ``<list>``.
 
     Args:
         node (ListBlock): The list IR node.
         parent (ET.Element): The ET element to attach the ``<list>`` to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_node`.
     """
     list_el = ET.SubElement(parent, _tei("list"))
     if node.ordered:
@@ -261,60 +283,64 @@ def _render_list_block(node: ListBlock, parent: ET.Element) -> None:
     for item in node.items:
         item_el = ET.SubElement(list_el, _tei("item"))
         for child in item.children:
-            _render_node(child, item_el)
+            _render_node(child, item_el, lookup)
 
 
-def _render_ref(node: Ref, parent: ET.Element) -> None:
+def _render_ref(node: Ref, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a :class:`~utils.nodes.Ref` IR node as a TEI ``<ref>``.
 
     Args:
         node (Ref): The hyperlink IR node.
         parent (ET.Element): The ET element to attach the ``<ref>`` to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_node`.
     """
     el = ET.SubElement(parent, _tei("ref"))
     el.set("target", node.target)
     for child in node.children:
-        _render_node(child, el)
+        _render_node(child, el, lookup)
 
 
-def _render_table(node: Table, parent: ET.Element) -> None:
+def _render_table(node: Table, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a :class:`~utils.nodes.Table` IR node as a TEI ``<table>``.
 
     Args:
         node (Table): The table IR node.
         parent (ET.Element): The ET element to attach the ``<table>`` to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_row`.
     """
     table_el = ET.SubElement(parent, _tei("table"))
     for row in node.rows:
-        _render_row(row, table_el)
+        _render_row(row, table_el, lookup)
 
 
-def _render_row(node: Row, parent: ET.Element) -> None:
+def _render_row(node: Row, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a :class:`~utils.nodes.Row` IR node as a TEI ``<row>``.
 
     Args:
         node (Row): The row IR node.
         parent (ET.Element): The ET element to attach the ``<row>`` to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_cell`.
     """
     row_el = ET.SubElement(parent, _tei("row"))
     if node.is_header:
         row_el.set("role", "label")
     for cell in node.cells:
-        _render_cell(cell, row_el)
+        _render_cell(cell, row_el, lookup)
 
 
-def _render_cell(node: Cell, parent: ET.Element) -> None:
+def _render_cell(node: Cell, parent: ET.Element, lookup: dict[int, Note]) -> None:
     """Render a :class:`~utils.nodes.Cell` IR node as a TEI ``<cell>``.
 
     Args:
         node (Cell): The cell IR node.
         parent (ET.Element): The ET element to attach the ``<cell>`` to.
+        lookup (dict[int, Note]): Forwarded to :func:`_render_node`.
     """
     cell_el = ET.SubElement(parent, _tei("cell"))
     if node.colspan is not None:
         cell_el.set("cols", str(node.colspan))
     for child in node.children:
-        _render_node(child, cell_el)
+        _render_node(child, cell_el, lookup)
 
 
 # ---------------------------------------------------------------------------
